moves home.jsx o the correct older and corrects import file paths
