+++ src/App.jsx
@@ -1,5 +1,6 @@
 import "./App.css";
 import Header from "./components/Header";
+import Home from "./pages/Home";
 import AdminLogin from "./pages/AdminLogin";
 import AdminDashboard from "./pages/AdminDashboard";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
@@ -9,8 +10,10 @@ export default function App() {
   return (
     <BrowserRouter>
       <Header />
-      <SuccessModal />
+      <Home />
+       <SuccessModal /> 
       <Routes>
+        <Route path="/" element={<Home/>} />
         <Route path="/admin-login" element={<AdminLogin />} />
         <Route path="/admin-dashboard" element={<AdminDashboard />} />
       </Routes>
