+++ src/main.jsx
@@ -2,8 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
-import { BrowserRouter } from "react-router-dom";
-import Header from "./components/Header";
+
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
